Inform users when there's an OTA update available
This commit causes the app to check:

1. On Launch
2. On Foreground

(so whenever the app goes from offscreen to onscreen) if there is an
over the air update available.

If there is an over the air update available, we try to download the
update.

If the download fails, nothing happens.

If the download succeeds, we let the user know that there is an update
available (and ready to be used) and ask them if we may restart the app
in order to apply the update.

The user can decline, in which case the update will be applied the next
time the app is restarted.