Prevent whitespace from staying on list screens
Previously we had an issue causing lists, when scrolled to the bottom,
to stay behind the iOS home indicator.

To solve that problem, we put an empty view at the bottom of the screen
(which always showed, regardless of the scroll position in the list) to
prevent the list from showing behind the iOS home indicator.

The problem with that solution is that the white view always shows,
causing the list to look "cut off" at the bottom, constantly.

We solve that problem in this commit by only showing the empty "footer-
padding" view when the list is scrolled all the way to the bottom.