Show game schedule
As a beer league hockey player,
When I open the Wolfpack app,
Then I see a tab titled "Games".

When I tap the tab titled "Games",
Then I see a list of the dates and times of my team's games,
So that I know when to go to the rink.

https://www.pivotaltracker.com/story/show/186975068