Prevent team names spanning multiple lines
Opposing team names are sometimes long and span multiple lines.

This commit ensures they do not take more than one line (if they try to,
because they're that long, they'll be truncated with an ellipses).

It also gives the React Native View containing those opposing team name
labels more width, in order to more likely fit the opposing team's name,
instead of truncating it with an ellipses.

It's also important that the two side columns of info, (when games are,
on the left - and the outcomes of games on the right) are the same
widths, so that the center column of information (where games are; their
rink and whether they're a home or away game) stays centered in the
screen.