Signup and login
This commit refactors signing up and logging in.

Previously, we only allowed administrators using Active Admin to create
players with phone numbers. Then, players could download the app and
verify themselves with their phone numbers.

Now, anyone can download the app, sign up and sign in.

However, there is no option to create a team, if you sign up and you
haven't already been created by an admin with your phone number.

This commit is a half-baked feature, for which we're ashamed.

It allows for anyone to potentially be able to use the app without
being pre-signed-up by an administrator but doesn't finish the job.

The reason this commit is being merged, as is, is because we started
work on it months ago and we'd really like to get it merged into the
main branch while we have all our checks passing (tsc is passing for
type checking, prettier validates, eslint validates, the tests pass, and
the test coverage is at 100%). We'd like to get this merged and move
forward with upgrading Expo (which has come out with a new SDK version
since this app was last worked on).

https://expo.dev/changelog/2024/05-07-sdk-51

This commit also replaces MSW with nock.

https://mswjs.io
https://github.com/nock/nock

MSW is "better" because it is network-request-making-technology (eg
fetch, axios, etc) independent. However, working with, and maintaining
a typescript-type-safe duplicate of of the rails' api's api with msw was
too time consuming.

There are a few factors that went into the decision to use nock over
msw. We wrote about them.

https://pachulski.me/choosing-an-api-mocking-library-to-use-with-jest