Update app name to Too Many Men
The app is intended to be used first by our hockey team, named Wolfpack.
However, the next few teams we'd like to get to start using it play at
the same rink that Wolfpack plays at and if the name of the app is
branded particularly just for the Wolfpack team, then we're afraid
others may not download and use this app only for that reason. Instead,
we're naming it Too Many Men instead of Wolfpack Hockey before putting
the app in the store.