Reselect the current team
Prior to this commit, after logging in, the player selects a team for
which they'd like to see the game schedule and the names of their
teammates. Once selected, there's no way to view the schedule or
teammate names for another team.

After this commit, a back button has been added which allows reselecting
the team that the player is currently viewing.