Fix deep links
Deep links were broken. We amended a couple tests which were already
reproducing the problem, and in a way ignoring the problem.

When a user is deep linked to, say `/teams/[teamId]`, which shows a list
of players, or to `/teams/[teamId]/games`, which shows of games, the
only component that is mounted is the Games or Players components. Both
of these components require there being a `teamId` in the URL. Clicking
the "back" button on either of those screens to go back to the team
selection screen (`/teams`) does not unmount them, since they're now the
root component in the navigation stack because it was the starting point
for the deep link. When that back button is pressed, since the
Games/Players components stays mount, and the URL loses the `teamId`
param, request are made to the `/teams/undefined/players` or
`/teams/undefined/games` endpoints of the rails API. That request
returns a 404 not found status and the sorting functionality in those
components (which assume the games or players are an array) to sort
players alphabetically by last name or games chronologically by their
played at times, breaks. This commit ensures that those requests aren't
made when the `teamId` url param is not present, so that their sorting
implementation does not crash the app.