Permit deep linking to players and games details
Deep linking to a players or games details screen causing the back,
provided by ExpoRouter.useRouter().back() to break when clicking the
back button in the navigation bar. It does nothing because there is no
route to go back to when you deep link directly into the details screen
of a game or player.

This fix permits deep linking into a games or players details screen
and gives the back button the functionality it needs (going back to the
games or players list (or "index") screens).

It looks like that (fixes the back button) by preloading the previous
index screen, which in turn fetches the data for that index screen), so
many tests were updated to mock the request for their index screen,
since the tests typically boot the app as if the player were deep linked
into that screen.

The fix is documented at the top of Expo's Expo Router FAQ. It looks
like changes may be made soon to improve this behavior. Hopefully, we'll
be able to revert this commit, which resulted in a big and unfortunately
not super clear / kind of confusing diff.

More info here: https://docs.expo.dev/router/reference/faq/#missing-back-button

From that page:

---

> If you set up a modal or another screen that is expected to have a back button, then you'll need to add unstable_settings to the route's layout to ensure the initial route is configured. Initial routes are somewhat unique to mobile apps and therefore fit awkwardly in the system — improvements pending.

```
// app/_layout.tsx

export const unstable_settings = {
  initialRouteName: 'index',
};
```