Keep players authenticated when they restart app
Once a user signs in, they may quit the app (if they're one of those
people...), restart their phone (causing the app to restart), or
receive one of the popups we send them when an Over-The-Air (OTA) update
is available which requires restarting the app to apply. In any of those
circumstances, we'd prefer not to force them to re-enter their phone
number and confirmation code.

Also... sending those confirmation codes with Twilio costs money. This
feature is a nice-to-have for users _and_ saves us money.