Show a loading spinner while auto signing in
Prior to this commit, when a user who has authenticated quits and
reopens the app, they are auto-reauthenticated, but the state
determining whether or not they get auto-reauthenticated is held in
async storage, which takes a non-negligable amount of time to access;
So, the user would see the login screen for a second when reopening the
app.

This commit shows a loading spinner while async storage is accessed
rather than flashing the login screen before signing the user back in.